should fix compilation error found on mips-irix in spectrum.cpp
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@5376 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/graphics/spectrum.cpp
@@ -2402,7 +2402,7 @@ Rebuilds the display_list for <spectrum> if it is not current.
 			}
 		}
 		table_size = (int) (number_of_texture_components *
-			 pow(number_of_values, number_of_data_components));
+			 pow((double) number_of_values, number_of_data_components));
 		if (ALLOCATE(colour_table, unsigned char, table_size))
 		{
 			colour_table_ptr = colour_table;
